requirements.txt
groq==0.12.0
ollama==0.3.3
pillow==11.0.0
PyAutoGUI==0.9.54
pyperclip==1.9.0
python-dotenv==1.0.1
